Replace callbacks with promises
And `LEVEL_NOT_FOUND` with `undefined`.

Passes the abstract test suite. Most other tests are commented out.

The `iterator.next()` method required a small change: previously we
had the callback signature `(err, entries, finished)` where
"finished" meant "reached the end of the data". Now we have a
promise for only `entries`. So an extra `next()` call (end-to-end)
is now needed to know that the end was reached.